requirements.txt
streamlit==1.52.2
torch==2.2.2
torchvision==0.17.2
Pillow==12.0.0
numpy==1.26.4
streamlit-drawable-canvas==0.9.3
opencv-python==4.12.0.88